EFFECTIVENESS OF “STUDENT-SERVICE INTERVENTION” TO LESSEN THE  TARDINESS OF SELECTED GRADE 12 STUDENTS
Abstract

The purpose of this study was to investigate the effectiveness of implementing Student-Service Intervention such as academic support services, academic advising or admissions in order to reduce the tardiness of Grade 12 students at San Pedro National High School in Bulacan. In the Philippines, being or starting things late has become a common issue, referred to as “Filipino Time”, with it having negative effects to students’ academic performance, behavior, and even overall well-being. A Quantitative experimental research design was used in this study and 20 students who were frequently late to class from each section were selected as the respondents. The researchers gathered the data through an intervention over a five-day period wherein the respondents were tasked to meet the researchers at a specific time in order to do activities such as cleaning and maintaining school facilities as the researchers list every time any of them were late. The study concluded that Student- Service Intervention was indeed effective in lessening the respondents’ tardiness and in enhancing student discipline.