ENHANCING THE ENGLISH-SPEAKING PROFICIENCY THROUGH “LOUD AND PROUD INTERVENTION” AMONG SELECTED GRADE 7 STUDENTS
Abstract

The English language is an essential skill to have as this is considered to be the universal language when it comes to socializing around the globe. With that this study aims to improve the English-speaking proficiency of students by assessing their vocabulary, grammar, and punctuation. This quantitative research used purposive sampling to select students that fit the criteria. The researchers used Storytelling, and Descriptive narrative of pictures to improve and develop the participants understanding. These students are given tasks to fill out the blanks, utilizing different unfamiliar words, and describing a picture's content where they are prohibited from using any language other than english to answer. The researchers concluded that the identified variables had significant effects on both speaking skills and english speaking skills, which shows that participants had been influenced from before and after.