EXAMINING THE INFLUENCE OF TEACHER-STUDENT RELATIONSHIP ON LEARNING BEHAVIOR OF GRADE 12 STUDENTS
Abstract
The main objective of this study is to determine the influence of the teacher-student relationship on the learning behavior of Grade 12 students. It aims to investigate whether the teacher's attitude affects the learning environment and its impact on the students' academic life. This qualitative study adopts a narrative approach. The researchers selected 30 final participants from San Pedro National High School using convenience sampling techniques. An interview questionnaire was developed by the researchers and administered to the participants. The findings of the study revealed that the teacher-student relationship significantly influences the learning behavior of Grade 12 students.