Angcajas, Shadie Shane M. Agulto, Hazel A. Bernardo, Sean Patrick Nagao, Aumi A. Puno, Rina Angelica M. Villanueva, Leonard
Abstract
It is clear that every student's decisions are influenced by the parents' decision-making. The
students, present career path is a result of those decisions. At this point, the researchers began to
investigate the potential effects of parental decision-making on the present career path of the students.
Purposive sampling was utilized by the researchers in this qualitative study to compile data from a total
of 55 respondents at San Pedro National High School. The researchers created a survey questionnaire
to ask participants for the data needed. The study's conclusions showed that parents' choices have a
impact on the present career paths that Technical-Vocational Livelihood (TVL) students are currently
thinking.