Calayag, Bea Loraine C. Manansala, Marianne M. Napoles, Carl Joemark M. Pascual, Angel M. Santos, Jhody Anne A.
Abstract
The main objective of this study is to determine the level of awareness and experiences
of selected grade 12 students on sexual inquisitiveness. This mix method study was applied
in this study specifically, the correlational and phenomenological design. The researchers
selected 55 respondents from San Pedro National High School using random sampling
technique. An interview guide questions and survey questions were developed and
administered to the participants. The result of this study shows that many of the students are
curious, aware and experienced on doing sexual activities.