Crisostomo, Eurika Kovy DR. Manalang, Hannah Mae C. Puzon, Robylyn S. Quizon, Christian Jay DL. Robles, Jayron G. Tanjuan, Maxine M.
Abstract
In the modern world, social media has become a huge part of our lives. The most frequent users
of it are in the new generation: millennials and Generation Z. Therefore, those are classified as students.
The objective of this research is to examine the social media usage of the students, particularly the
TikTok application. This study is useful for providing insights into the TikTok application usage among
Senior High School students, especially in the field of audio-visual communication. To be able to
achieve the said goal, the researchers conducted a survey with the selected grade 12 respondents. The
questionnaire consists of questions that describe their TikTok usage. The data analysis revealed that the
majority of respondents use TikTok primarily for video watching. The primary activity on TikTok is
content consumption rather than purchasing or publishing. It is followed by students who are not TikTok
users or simply are not using the TikTok application.The implications of this study could be used to
raise concerns about what people, particularly students, do the majority of the time when using social
media, specifically Tiktok.