Atienza, Roslyn L. Castro, Jham Fritz M. Inocencio, Aaron James DC. Locsin, Decie Claire A. Reyes, Ryan Kenneth F.
Abstract
This study aims to know the impacts of parenting styles to the academic achievement and
behavior of selected Grade 9 students in San Pedro National High School. This study was quantitative
in nature. The researchers chose 200 selected students as their respondents and simple random
sampling technique was used to select them. The survey questionnaire used in this study contains 25
items that were adopted from the studies of Ofosu (2013), Amponsah et. al. (2018), and Munyi (2013).
As a conclusion, the researchers found out that parenting styles have significant effect to the academic
achievement and behavior of the students.